Animation hack based on observations from https://github.com/jpsim/UICollectionView-Animation-Bug
We make the collection view larger than the rendered area, spilling over the top and bottom of the screen.  That way when the layout is invalidated the collection view will have a view to move to the new location rather than just instantly appearing/disappearing it.  This has a performance penalty as additional views are instantiated and flung about so perform appropriate optimizations.